+++ firebaseConfig.js
@@ -24,4 +24,4 @@ export const auth = initializeAuth(app, {
 });
 
 // Initialize Cloud Firestore and get a reference to the service
-const db = getFirestore(app);
+export const db = getFirestore(app);
